Move itemized prompts into separate module (#4229)
* Move itemized prompts into separate module

* Update public/scripts/itemized-prompts.js


* Update public/scripts/itemized-prompts.js


---------
